Modification on gg_rescale.py
Added in the ability to rescale the model with an input mean molecular weight scaling for the atmosphere. Changed the variable names from two letter variables to descriptive words. Added in a second run in the main: to show how the mean molecular weight scaling compares to the default of a H/He-dominated atmosphere.